+++ src/index.ts
@@ -1,7 +1,16 @@
-export { each } from './each.js'
-export { map } from './map.js'
+export { each, EachOptions } from './each.js'
+export { map, MapOptions } from './map.js'
 export { Deferred, createDeferred } from './deferred.js'
 export { gowait, ErrorValue } from './gowait.js'
 export { Semaphore, createLock } from './semaphore.js'
-export { Queue, q, QueueOptions, QueueSettleMode } from './queue.js'
+export {
+  AbortError,
+  Queue,
+  q,
+  QueueOptions,
+  QueueSettleMode,
+  QueueTaskOptions,
+  QueueWaitOptions,
+  TimeoutError,
+} from './queue.js'
 export { Func, noop, identity, once, delay } from './lang.js'